Tex now returns the generated TeX code as a string.
Not sure where the responsibilities lie.  If TeX returns a string
then the caller can decide what to do with the string.  However,
large amounts of TeX code will requre large amounts of memory if
we return a string rather than writing straight to a file (for
exmaple).  For now, assume that large amounts of TeX will never
get generated.